requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
openai==1.30.1
jinja2==3.1.3
python-dotenv==1.0.1
aiofiles==23.2.1
